+++ src/hassette/config/core.py
@@ -153,6 +153,11 @@ class HassetteConfig(BaseSettings):
     app_shutdown_timeout_seconds: int = Field(default=10)
     """Length of time to wait for an app to shut down before giving up."""
 
+    resource_shutdown_timeout_seconds: int = Field(
+        default_factory=lambda data: data.get("app_shutdown_timeout_seconds", 10)
+    )
+    """Length of time to wait for a resource to shut down before giving up. Defaults to app_shutdown_timeout_seconds."""
+
     websocket_authentication_timeout_seconds: int = Field(default=10)
     """Length of time to wait for WebSocket authentication to complete."""
 
@@ -197,6 +202,9 @@ class HassetteConfig(BaseSettings):
 
     task_cancellation_timeout_seconds: int = Field(default=5)
     """Length of time to wait for tasks to cancel before forcing."""
+
+    default_cache_size: int = Field(default=100 * 1024 * 1024)
+    """Default size limit for caches in bytes. Defaults to 100 MiB."""
 
     # Service log levels