+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `diskcache` dependency and `cache` attribute to all resources
-  - Each resource (based on `class_name`) gets its own disk-backed cache instance
+  - Each resource class has its own cache directory under the Hassette data directory
 
 ## [0.16.0] - 2025-11-16