+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.17.0] - 2025-11-22
+
 ### Changed
 - **Breaking:** - Requires Python 3.13 going forward, Python 3.12 and 3.11 are no longer supported.
   - This allows use of `type`, defaults for TypeVars, and other new typing features.